services/dispatcher/requirements.txt
boto3==1.34.0
psycopg2-binary==2.9.9
requests==2.31.0
pytz==2024.1
alpaca-py==0.43.2
